[CID-1842] - additional test

+++ src/nonCommitTest/groovy/com/rackspace/idm/api/resource/cloud/v20/UserManagerAccessibleResourcesIntegrationTest.groovy
@@ -1,12 +1,18 @@
 package com.rackspace.idm.api.resource.cloud.v20
 
+import com.rackspace.docs.identity.api.ext.rax_auth.v1.Domains
 import com.rackspace.docs.identity.api.ext.rax_kskey.v1.ApiKeyCredentials
+import com.rackspace.idm.Constants
 import org.apache.commons.lang3.StringUtils
 import org.openstack.docs.identity.api.v2.AuthenticateResponse
+import org.openstack.docs.identity.api.v2.CredentialListType
+import org.openstack.docs.identity.api.v2.EndpointList
 import org.openstack.docs.identity.api.v2.Tenant
 import org.openstack.docs.identity.api.v2.User
+import org.openstack.docs.identity.api.v2.UserList
 import spock.lang.Shared
 import testHelpers.RootIntegrationTest
+import testHelpers.saml.SamlFactory
 
 import static com.rackspace.idm.Constants.*
 import static org.apache.http.HttpStatus.*
@@ -14,10 +20,16 @@ import static org.apache.http.HttpStatus.*
 class UserManagerAccessibleResourcesIntegrationTest extends RootIntegrationTest {
 
     @Shared
+    String domainId
+
+    @Shared
     User userAdmin, userManage, userManage2
 
     @Shared
-    String userManageToken, userAdminToken, identityAdminToken
+    def fedUser
+
+    @Shared
+    String userManageToken, userAdminToken, identityAdminToken, fedUserToken
 
     @Shared
     Tenant tenant
@@ -26,7 +38,7 @@ class UserManagerAccessibleResourcesIntegrationTest extends RootIntegrationTest 
         identityAdminToken = cloud20.authenticatePassword("auth", "auth123").getEntity(AuthenticateResponse).value.token.id
 
         def random = UUID.randomUUID()
-        def domainId = "domainId$random"
+        domainId = "domainId$random"
 
         // User Admin
         def response = cloud20.createUser(identityAdminToken, v2Factory.createUserForCreate("userAdmin$random", "display", "test@rackspace.com", true, null, domainId, DEFAULT_PASSWORD))
@@ -51,6 +63,7 @@ class UserManagerAccessibleResourcesIntegrationTest extends RootIntegrationTest 
         assert response.status == SC_CREATED
 
         tenant = response.getEntity(Tenant).value
+
     }
 
     def cleanupSpec() {
@@ -63,10 +76,21 @@ class UserManagerAccessibleResourcesIntegrationTest extends RootIntegrationTest 
         // Create second userManager for testing
         userManage2 = utils.createUser(utils.getToken(userAdmin.username))
         utils.addRoleToUser(userManage2, USER_MANAGE_ROLE_ID)
+
+        // Create federated user with user-manage role
+        def username = testUtils.getRandomUUID("userForSaml")
+        def samlAssertion = new SamlFactory().generateSamlAssertionStringForFederatedUser(DEFAULT_IDP_URI, username, DEFAULT_SAML_EXP_SECS, domainId, [Constants.USER_MANAGE_ROLE_NAME], "fedUser@mail.com");
+        def samlResponse = cloud20.samlAuthenticate(samlAssertion)
+        AuthenticateResponse authResponse = samlResponse.getEntity(AuthenticateResponse).value
+        fedUser = authResponse.user
+        assert authResponse.user.roles.role.find {it.id == USER_MANAGE_ROLE_ID} != null
+
+        fedUserToken = authResponse.token.id
     }
 
     def cleanup() {
         utils.deleteUserQuietly(userManage2)
+        utils.deleteUserQuietly(fedUser)
     }
 
     def "user-manage can add/delete role to another user-manage in the same domain"() {
@@ -78,6 +102,18 @@ class UserManagerAccessibleResourcesIntegrationTest extends RootIntegrationTest 
 
         when: "delete role"
         def deleteResponse = cloud20.deleteApplicationRoleFromUser(userManageToken, ROLE_RBAC1_ID, userManage2.id)
+
+        then:
+        deleteResponse.status == SC_NO_CONTENT
+
+        when: "add role - federated user caller"
+        addResponse = cloud20.addUserRole(fedUserToken, userManage2.id, ROLE_RBAC1_ID)
+
+        then:
+        addResponse.status == SC_OK
+
+        when: "delete role - federated user caller"
+        deleteResponse = cloud20.deleteApplicationRoleFromUser(fedUserToken, ROLE_RBAC1_ID, userManage2.id)
 
         then:
         deleteResponse.status == SC_NO_CONTENT
@@ -102,6 +138,18 @@ class UserManagerAccessibleResourcesIntegrationTest extends RootIntegrationTest 
         then:
         deleteResponse.status == SC_FORBIDDEN
 
+        when: "add role - federated user caller"
+        addResponse = cloud20.addUserRole(fedUserToken, userManageOther.id, ROLE_RBAC1_ID)
+
+        then:
+        addResponse.status == SC_FORBIDDEN
+
+        when: "delete role - federated user caller"
+        deleteResponse = cloud20.deleteApplicationRoleFromUser(fedUserToken, ROLE_RBAC1_ID, userManageOther.id)
+
+        then:
+        deleteResponse.status == SC_FORBIDDEN
+
         cleanup:
         utils.deleteUsersQuietly(users)
     }
@@ -115,6 +163,18 @@ class UserManagerAccessibleResourcesIntegrationTest extends RootIntegrationTest 
 
         when: "delete role from user on tenant"
         def deleteResponse = cloud20.deleteRoleFromUserOnTenant(userManageToken, tenant.id, userManage2.id, ROLE_RBAC1_ID)
+
+        then:
+        deleteResponse.status == SC_NO_CONTENT
+
+        when: "add role to user on tenant - federated user caller"
+        response = cloud20.addRoleToUserOnTenant(fedUserToken, tenant.id, userManage2.id, ROLE_RBAC1_ID)
+
+        then:
+        response.status == SC_OK
+
+        when: "delete role from user on tenant - federated user caller"
+        deleteResponse = cloud20.deleteRoleFromUserOnTenant(fedUserToken, tenant.id, userManage2.id, ROLE_RBAC1_ID)
 
         then:
         deleteResponse.status == SC_NO_CONTENT
@@ -135,6 +195,18 @@ class UserManagerAccessibleResourcesIntegrationTest extends RootIntegrationTest 
 
         when: "delete role from user on tenant"
         def deleteResponse = cloud20.deleteRoleFromUserOnTenant(userManageToken, tenant.id, userManageOther.id, ROLE_RBAC1_ID)
+
+        then:
+        deleteResponse.status == SC_FORBIDDEN
+
+        when: "add role to user on tenant - federated user caller"
+        response = cloud20.addRoleToUserOnTenant(fedUserToken, tenant.id, userManageOther.id, ROLE_RBAC1_ID)
+
+        then:
+        response.status == SC_FORBIDDEN
+
+        when: "delete role from user on tenant - federated user caller"
+        deleteResponse = cloud20.deleteRoleFromUserOnTenant(fedUserToken, tenant.id, userManageOther.id, ROLE_RBAC1_ID)
 
         then:
         deleteResponse.status == SC_FORBIDDEN
@@ -173,6 +245,12 @@ class UserManagerAccessibleResourcesIntegrationTest extends RootIntegrationTest 
         then:
         response.status == SC_FORBIDDEN
 
+        when: "update user's email - federated user caller"
+        response = cloud20.updateUser(fedUserToken, userManageOther.id, userToUpdate)
+
+        then:
+        response.status == SC_FORBIDDEN
+
         cleanup:
         utils.deleteUsersQuietly(users)
     }
@@ -195,10 +273,25 @@ class UserManagerAccessibleResourcesIntegrationTest extends RootIntegrationTest 
 
         then:
         resetResponse.status == SC_OK
-        assert StringUtils.isNoneBlank(resetCred.apiKey)
+        assert StringUtils.isNotBlank(resetCred.apiKey)
+
+        when: "delete API key - federated user caller"
+        response = cloud20.deleteUserApiKey(fedUserToken, userManage2.id)
+
+        then:
+        response.status == SC_NO_CONTENT
+
+        when: "reset and get API key - federated user caller"
+        resetResponse = cloud20.resetUserApiKey(fedUserToken, userManage2.id)
+        resetCred = utils.getUserApiKey(userManage2)
+
+        then:
+        resetResponse.status == SC_OK
+        assert StringUtils.isNotBlank(resetCred.apiKey)
     }
 
     def "error check: user-manage CANNOT get/delete/reset another user-manage API key credentials in a different domain"() {
+        given:
         def domainIdOther = utils.createDomain()
         def identityAdminOther, userAdminOther, userManageOther, defaultUserOther
         (identityAdminOther, userAdminOther, userManageOther, defaultUserOther) = utils.createUsers(domainIdOther)
@@ -227,4 +320,135 @@ class UserManagerAccessibleResourcesIntegrationTest extends RootIntegrationTest 
         utils.deleteUsersQuietly(users)
     }
 
+    def "federated user with user-manager can view user-managers within same domain"() {
+        when: "get user by id - self"
+        def response = cloud20.getUserById(fedUserToken, fedUser.id)
+        User user = response.getEntity(User).value
+
+        then:
+        response.status == SC_OK
+
+        user.id == fedUser.id
+
+        when: "get user by id - other user-manager"
+        response = cloud20.getUserById(fedUserToken, userManage.id)
+        user = response.getEntity(User).value
+
+        then:
+        response.status == SC_OK
+
+        user.id == userManage.id
+
+        when: "get user by name"
+        response = cloud20.getUserByName(fedUserToken, userManage.username)
+        user = response.getEntity(User).value
+
+        then:
+        response.status == SC_OK
+
+        user.id == userManage.id
+
+        when: "get user by email"
+        response = cloud20.getUsersByEmail(fedUserToken, userManage.email)
+        UserList userList = response.getEntity(UserList).value
+
+        then:
+        response.status == SC_OK
+
+        userList.user.size() == 1
+        userList.user.find {it.id == userManage.id}
+
+        when: "list users"
+        response = cloud20.listUsers(fedUserToken)
+        userList = response.getEntity(UserList).value
+
+        then:
+        response.status == SC_OK
+
+        userList.user.size() == 3
+        userList.user.find {it.id == userManage.id} != null
+        userList.user.find {it.id == userManage2.id} != null
+        userList.user.find {it.id == fedUser.id} != null
+    }
+
+    def "error check: federated user with user-manager"() {
+        given:
+        def domainIdOther = utils.createDomain()
+        def identityAdminOther, userAdminOther, userManageOther, defaultUserOther
+        (identityAdminOther, userAdminOther, userManageOther, defaultUserOther) = utils.createUsers(domainIdOther)
+        def users = [defaultUserOther, userManageOther, userAdminOther, identityAdminOther]
+
+        when: "get user by id - user-manager in different domain"
+        def response = cloud20.getUserById(fedUserToken, userManageOther.id)
+
+        then:
+        response.status == SC_NOT_FOUND
+
+        when: "get user by name - user-manager in different domain"
+        response = cloud20.getUserByName(fedUserToken, userManageOther.username)
+
+        then:
+        response.status == SC_FORBIDDEN
+
+        when: "get user by email"
+        response = cloud20.getUsersByEmail(fedUserToken, userManageOther.email)
+        UserList userList = response.getEntity(UserList).value
+
+        then:
+        response.status == SC_OK
+
+        userList.user.size() == 0
+
+        cleanup:
+        utils.deleteUsersQuietly(users)
+    }
+
+    def "federated user with user-manager can do accessible domain operations on user-manager in same domain"() {
+        when: "get accessible domains for user"
+        def response = cloud20.getAccessibleDomainsForUser(fedUserToken, userManage.id)
+        Domains domains = response.getEntity(Domains)
+
+        then:
+        response.status == SC_OK
+
+        domains.domain.size() == 1
+        domains.domain.find {it.id == domainId}
+
+        when: "get accessible domain endpoints for user"
+        response = cloud20.getAccessibleDomainEndpointsForUser(fedUserToken, userManage.id, domainId)
+        EndpointList endpointList = response.getEntity(EndpointList).value
+
+        then:
+        response.status == SC_OK
+        endpointList != null
+    }
+
+    def "error check: federated user with user-manager CANNOT do accessible domain operations on user-manager in different domain"() {
+        given:
+        def domainIdOther = utils.createDomain()
+        def identityAdminOther, userAdminOther, userManageOther, defaultUserOther
+        (identityAdminOther, userAdminOther, userManageOther, defaultUserOther) = utils.createUsers(domainIdOther)
+        def users = [defaultUserOther, userManageOther, userAdminOther, identityAdminOther]
+
+        when: "get accessible domains for user - self"
+        def response = cloud20.getAccessibleDomainsForUser(fedUserToken, fedUser.id)
+
+        then:
+        response.status == SC_NOT_FOUND // Only works for provisioned users
+
+        when: "get accessible domains for user"
+        response = cloud20.getAccessibleDomainsForUser(fedUserToken, userManageOther.id)
+
+        then:
+        response.status == SC_FORBIDDEN
+
+        when: "get accessible domain endpoints for user"
+        response = cloud20.getAccessibleDomainEndpointsForUser(fedUserToken, userManageOther.id, domainIdOther)
+
+        then:
+        response.status == SC_FORBIDDEN
+
+        cleanup:
+        utils.deleteUsersQuietly(users)
+    }
 }
